Update seedinternalusers command to accept team and default queue
The users created using this command are by default belong to Admin
team and queue is set to 'All cases'. This is a problem in E2E tests
as when we sign in as different users their team and queue are not
correct. Because of this we are having to switch team, queue as
part of the test and as these persist between tests it can cause
test failures.

Better approach is to seed certain users with expected roles,
teams, queues etc.