Add migration to back populate LicenceDecision instances
These are created for issued, refused and revoked outcomes.

We use CREATE_FINAL_RECOMMENDATION timestamp as the primary source when
back populating licence decisions however this audit log is not available
from the beginning. So for the case where this log is missing we
just take the licence decision document creation date as the
decision date.