Update migration to populate with the actual licence at that time
When a new licence decision is created we are now going to populate
licence also. This was missing in the back populated data.

Previously we added this migration to back populate but it was done
to match with report data which would populate with the earliest
licence for that decision but in reality that would not be the case.

Update migration to populate with actual licence at that time.