Add previous decision to Licence decision to save previous version of this decision
A case can have multiple licence decisions (outcomes) and when reporting
if the decision is same in all cases we report on the initial decision date.

If the decision changes then we still report on the initial decision date but
also subsequent changes.

To achieve this effectively without adding too many fields a self referencing
field 'previous_decision' is added to this model. This is populated such that
it gets reset whenever there is a change in decision otherwise points to previous
decision. When exporting data this is used to filter decisions we want to
ignore and pick the correct decision date.