Update migration to populate licence field correctly
Currently we have changes in live that populate licence field in
LicenceDecision when a decision is made. Using the previous approach
the timestamps were not matching for those newly created decisions.

This can be simplified a lot by ignore all objects where licence
is set and the timestamps for the remaining ones. For these objects
use audit log and documents to populate the licence.