Fix sorting issue
Current code sorting the results based on first key in the results but for
one of the test this is same for all rows so randomly it sorts by different
order and causing the test to fail. Update the test to sort by all keys
so we have a consistent order.