Add unit test for the move case forward function
The test takes a case that is assigned to an OGD queue and moves it forward.
To be able to verify that the current queue is unassigned and next queue
as per routing rules is assigned correctly the case is initially processed
through product assessments and pre-circ. This is to prevent triggering of
fallback routing rules which otherwise doesn't unassign the current queue
and in addition routes to this fallback queue as this rules relies on
presence of specific audit events.

Though this verifies next queue as per routing rules this is not be
considered a test for routing rules.