Create StandardSubmittedApplicationFactory for flags tests
For these tests we want an application to be in submitted state
without actually submitting (via API call) as that would trigger
flagging rules and alter the flags applied.

We cannot use DraftStandardApplicationFactory() as its purpose
is to give a draft application. StandardApplicationFactory() can
be updated but no products attached at this stage.

A new factory is created for this purpose which has products and
provided flags attached.

When changes are saved then post_save signal can still trigger
flagging rules hence this is mocked in the test.