Record post-circ queue exit date after finalising case
When a case is finalised from post-circ queue then the exit date
is not recorded as we don't move it forward (unless for countersigning).

Hence we need to handle this after finalising to record exit date.
Even in case of LU countersigining as the case comes back to this
queue for finalising after all countersignatures.