Update migration that populates user in CaseQueueMovement
Previously in the migration we were iterating through each object
and updating the 'user' field which was timing out.

Looking at all MOVE_CASE events in UAT, all of the users are SystemUser
so we can just update user for all instances of CaseQueueMovement
as SystemUser without iterating through MOVE_CASE events.