Update CaseCanUserMakeRecommendations permission
Previously this permission fails if same user tries to post recommendations
again on a case but now the OGD recommendation flow is updated to
select multiple entities at a time. User may either select all or
subset of entities at a time and he will have to post multiple times
to make recommendations on all entities.

So remove this check from this permission and allow the user to post
multiple times.